Addresses #2: Path parameters not supported
The user does not have the ability to add a path parameter to the
swagger doc. Instead, on config open, it checks to see if there are path
parameter docs defined for each path param in the url. If not, it adds
one. These path parameters can only have the description and type
changed. They cannot be removed. Instead, the editor also checks and
cleans up old path parameters that no longer exist.